ai-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
torch==2.2.1
torchvision==0.17.1
Pillow==10.2.0
numpy<2.0.0
opencv-python-headless==4.9.0.80
python-multipart==0.0.9
httpx==0.27.0
pytest==8.1.1
pytest-asyncio==0.23.5
torchxrayvision
scikit-learn==1.4.1.post1
